api13-draft: fix real API13 source drift, now builds clean
Three genuine source-level gaps beyond the earlier vendor re-pins, found by
diffing D:\Dalamud's own FFXIVClientStructs/Dalamud source against what the
reverted (pre-TC-port) code assumed:

- MobPulled.cs: IClientState.TerritoryChanged's delegate parameter is ushort,
  not uint (Action<ushort> in real API13's IClientState.cs).
- ApproachingMapFlag.cs: AgentMap's FlagMapMarker/IsFlagMarkerSet moved from
  properties to ECommons.CSExtensions.AgentMapExtensions extension methods
  (AgentMap's _flagMapMarkers field is a FixedSizeArray-generated member in
  this generation, no longer a plain struct field) - needs () to invoke.
- FishBite.cs: confirmed via reflection against the real FFXIVClientStructs.dll
  that FishingEventHandler is still just a generic EventHandler* with no typed
  State field and no FishingState enum anywhere in the reference assemblies -
  same genuine gap TC's original port already found and stubbed out, restored
  verbatim (GetFishingState() always returns None; fish-bite notifications are
  a known no-op pending upstream FFXIVClientStructs support).

+++ NotificationMaster/Notificators/MobPulled.cs
@@ -55,7 +55,7 @@ internal class MobPulled : IDisposable
         PluginLog.Debug("Cleared ignored mobs ids cache");
     }
 
-    internal void TerritoryChanged(uint newTerritory)
+    internal void TerritoryChanged(ushort newTerritory)
     {
         Svc.Framework.Update -= MobPulledWatcher;
         PluginLog.Debug("MobPulledWatcher unregistered.");
